refactor(unity-ipc): split method dispatch handlers via DI
Why:
- UnityIpcMethodDispatcher had low cohesion because it contained all method-specific behaviors in one class.
- Method expansion should be additive through DI registration, not by growing dispatcher internals.

What Changed:
- Added IUnityIpcMethodHandler contract and split ping/execute/test.run/shutdown handlers into dedicated files.
- Changed UnityIpcMethodDispatcher to consume IEnumerable<IUnityIpcMethodHandler> and build method map with duplicate/invalid registration checks.
- Updated UnityDaemonBootstrap to register each method handler in DI.

Impact:
- Dispatcher now focuses only on routing and common error policy.
- Adding a new IPC method no longer requires editing dispatcher internals.
