+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Ipc/Server/IUnityIpcMethodHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 43a6ba30918484a169f1e7fcc14b88af
