+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Ipc/Server/PingUnityIpcMethodHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 600b4c8e3aacb49ae9c9161a337d3ede
