+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Ipc/Server/ShutdownUnityIpcMethodHandler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 9dca59c6b73e24cf8b60afed850926a6
