+++ src/Ucli.Unity/Assets/MackySoft/MackySoft.Ucli.Unity/Editor/Ipc/Server/TestRunUnityIpcMethodHandler.cs
@@ -1,0 +1,66 @@
+using System;
+using System.Threading;
+using System.Threading.Tasks;
+using MackySoft.Ucli.Contracts.Ipc;
+
+namespace MackySoft.Ucli.Unity.Ipc
+{
+    /// <summary> Handles <c>test.run</c> IPC method requests. </summary>
+    internal sealed class TestRunUnityIpcMethodHandler : IUnityIpcMethodHandler
+    {
+        private readonly IUnityTestRunService testRunService;
+
+        /// <summary> Initializes a new instance of the <see cref="TestRunUnityIpcMethodHandler" /> class. </summary>
+        /// <param name="testRunService"> The test-run service dependency. </param>
+        public TestRunUnityIpcMethodHandler (IUnityTestRunService testRunService)
+        {
+            this.testRunService = testRunService ?? throw new ArgumentNullException(nameof(testRunService));
+        }
+
+        /// <inheritdoc />
+        public string Method => IpcMethodNames.TestRun;
+
+        /// <inheritdoc />
+        public async ValueTask<IpcResponse> Handle (
+            IpcRequest request,
+            CancellationToken cancellationToken)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
+            ArgumentNullException.ThrowIfNull(request);
+
+            if (!UnityIpcRequestCodec.TryDecodeTestRunRequest(
+                    request,
+                    out IpcTestRunRequest? testRunRequest,
+                    out var errorResponse))
+            {
+                return errorResponse!;
+            }
+
+            try
+            {
+                var payload = await testRunService.Execute(testRunRequest!, cancellationToken);
+                return UnityIpcResponseFactory.CreateSuccessResponse(request, payload);
+            }
+            catch (OperationCanceledException)
+            {
+                throw;
+            }
+            catch (ArgumentException exception)
+            {
+                return UnityIpcResponseFactory.CreateErrorResponse(
+                    request,
+                    IpcErrorCodes.InvalidArgument,
+                    exception.Message,
+                    null);
+            }
+            catch (Exception exception)
+            {
+                return UnityIpcResponseFactory.CreateErrorResponse(
+                    request,
+                    IpcErrorCodes.InternalError,
+                    $"Unity test run failed. {exception.Message}",
+                    null);
+            }
+        }
+    }
+}
